+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ This document describes changes between each past release.
 - Now fully rely on Pyup.io (or contributors) to update the versions in the `requirements.txt` file (fixes #1512)
 - Move from importing pip to running it in a subprocess (see https://github.com/pypa/pip/issues/5081).
 - Remove useless print when using the OpenID policy (ref #1509)
+- Try to recover from the race condition where two requests can delete the same record. (Fix #1557; refs #1407.)
 
 
 8.2.0 (2018-03-01)